Fixed wrong merge, PerformanceApp is again available
